refactor: applying part of pr comments
