+++ src/components/Accordion/index.ts
@@ -1,1 +1,1 @@
-export { Accordion } from './Accordion';
+export { AccordionContainer } from './AccordionContainer';
