+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export { Accordion } from './Accordion';
+export { AccordionContainer } from './Accordion';
 export { Backdrop } from './Backdrop';
 export { Card } from './Card';
 export { Carousel } from './Carousel';
